+++ lib/qsim/impls/ansatz_generator/trotter.cpp
@@ -1,0 +1,26 @@
+#include "trotter.hpp"
+#include "xacc_service.hpp"
+
+namespace qcor {
+namespace qsim {
+Ansatz TrotterEvolution::create_ansatz(Observable *obs,
+                                       const HeterogeneousMap &params) {
+  Ansatz result;
+  // This ansatz generator requires an observable.
+  assert(obs != nullptr);
+  double dt = 1.0;
+  if (params.keyExists<double>("dt")) {
+    dt = params.get<double>("dt");
+  }
+  // Just use exp_i_theta for now
+  // TODO: formalize a standard library kernel for this.
+  auto expCirc = std::dynamic_pointer_cast<xacc::quantum::Circuit>(
+      xacc::getService<xacc::Instruction>("exp_i_theta"));
+  expCirc->expand({{"pauli", obs->toString()}});
+  result.circuit = expCirc->operator()({dt});
+  result.nb_qubits = expCirc->nRequiredBits();
+
+  return result;
+}
+} // namespace qsim
+} // namespace qcor
