+++ lib/qsim/impls/cost_evaluator/tests/resources/ibm_depol_1e-3_noise_model.json
