+++ lib/qsim/impls/cost_evaluator/tests/resources/ibmqx2_noise_model.json
