+++ lib/qsim/impls/workflow/qite.cpp
@@ -1,0 +1,16 @@
+#include "qite.hpp"
+
+namespace qcor {
+namespace qsim {
+bool QiteWorkflow::initialize(const HeterogeneousMap &params) {
+  // TODO
+  return true;
+}
+
+QuantumSimulationResult
+QiteWorkflow::execute(const QuantumSimulationModel &model) {
+  // TODO
+  return {};
+}
+} // namespace qsim
+} // namespace qcor
